requirements.txt

flask==3.0.0
gunicorn==20.1.0
chromadb==0.6.2
anthropic==0.7.7
python-dotenv==1.0.0
html2text==2024.2.26
psycopg2-binary
sqlalchemy
duckdb
